[dev]向上取整需要先转换为double数据类型

+++ src/main/java/com/itstyle/control/CarNumController.java
@@ -75,7 +75,7 @@ public class CarNumController {
         accessTypes = accessTypes.stream().filter(e -> e.getChannelTypeId() % 2 == 1).collect(Collectors.toList());
         model.addAttribute("accessTypes", accessTypes);
 
-        model.addAttribute("maxPage", Math.ceil(pageResponse.getCount() / 4));
+        model.addAttribute("maxPage", Math.ceil((double)(pageResponse.getCount() / 4)));
         return "/backend/tempcarinfo";
     }
